fixing som bugs, dividing create view to prevent to many components rerender
