+++ src/components/StrategyInterface/StrategyInterface.module.scss
@@ -1,0 +1,5 @@
+.strategyBarWrapper {
+  &:hover {
+    cursor: pointer;
+  }
+}
