+++ src/components/StrategyInterface/StrategyInterfaceElement/StrategyInterfaceElement.module.scss
@@ -1,6 +1,7 @@
 .element {
   position: relative;
   padding-left: 12px;
+  width: fit-content;
   &::before {
     content: "";
     position: absolute;
