+++ src/views/CreateView/CreateView.module.scss
@@ -26,9 +26,3 @@
   flex-direction: column;
   gap: 20px;
 }
-
-.strategyBarWrapper {
-  &:hover {
-    cursor: pointer;
-  }
-}
