--- src/views/MainView.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-const MainView = () => <></>;
-
-export default MainView;
